Allow for download on the review page
